Remove old storage mechanism (DataUtils, OngoingTripStorageHelper and StoredTripHelper)
