+++ android/app/src/main/java/edu/berkeley/eecs/cfc_tracker/BootReceiver.java
@@ -3,12 +3,6 @@ package edu.berkeley.eecs.cfc_tracker;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
-import android.preference.PreferenceManager;
-
-import java.io.IOException;
-
-import edu.berkeley.eecs.cfc_tracker.storage.DataUtils;
 
 /*
  * Class that allows us to re-register the alarms when the phone is rebooted.
@@ -27,7 +21,7 @@ public class BootReceiver extends BroadcastReceiver {
              * point and invalidate all entries in the database.
              * TODO: Remove this if we decide to switch back to utc time.
              */
-            DataUtils.endTrip(ctx);
+            // DataUtils.endTrip(ctx);
 
             // Re-initialize the state machine
             ctx.sendBroadcast(new Intent(ctx.getString(R.string.transition_initialize)));
